+++ libs/datamodel/core/src/lib.rs
@@ -12,7 +12,7 @@ pub mod configuration;
 pub mod dml;
 pub mod error;
 pub mod json;
-pub mod validator;
+pub mod transform;
 pub mod walkers;
 
 pub use common::DefaultNames;
@@ -21,7 +21,7 @@ pub use dml::*;
 
 use crate::ast::SchemaAst;
 use std::io::Write;
-use validator::ast_to_dml::ValidationPipeline;
+use transform::{ast_to_dml::ValidationPipeline, dml_to_ast::LowerDmlToAst};
 
 /// Parses and validates a datamodel string, using core attributes only.
 pub fn parse_datamodel(datamodel_string: &str) -> Result<Datamodel, error::ErrorCollection> {
@@ -156,7 +156,7 @@ pub fn render_datamodel_to(
     stream: &mut dyn std::io::Write,
     datamodel: &dml::Datamodel,
 ) -> Result<(), error::ErrorCollection> {
-    let lowered = validator::dml_to_ast::LowerDmlToAst::new().lower(datamodel)?;
+    let lowered = LowerDmlToAst::new().lower(datamodel)?;
     render_schema_ast_to(stream, &lowered, 2);
     Ok(())
 }
@@ -177,7 +177,7 @@ pub fn render_datamodel_and_config_to(
     datamodel: &dml::Datamodel,
     config: &configuration::Configuration,
 ) -> Result<(), error::ErrorCollection> {
-    let mut lowered = validator::dml_to_ast::LowerDmlToAst::new().lower(datamodel)?;
+    let mut lowered = LowerDmlToAst::new().lower(datamodel)?;
 
     SourceSerializer::add_sources_to_ast(config.datasources.as_slice(), &mut lowered);
     GeneratorLoader::add_generators_to_ast(&config.generators, &mut lowered);
